backend/requirements.txt
fastapi
uvicorn[standard]
letta
httpx
psutil
GPUtil
python-dotenv
